karajo: HTTP workers and manager with web user interface
Module karajo implement HTTP workers and manager similar to AppEngine
cron.

karajo has the web user interface (WUI) for monitoring the jobs that run
on port 31937 by default and can be configurable.

A single instance of karajo is configured through an Environment or
loaded from ini file format. There are three configuration sections,
one to configure the server, one to configure the logs, and another one
to configure one or more jobs to be executed.

For more information see the README file in this repository.