all: merge the LogOptions into Environment
The LogOptions.Dir become Environment.DirLogs.

The LogOptions.FilenamePrefix become Environment.Name.
The Name field will be used for title on the web user interface, as log
prefix, for file prefix on the jobs state, and as file prefix on
log files.  If this value is empty, it will be set to "karajo".