www: make the showAttrs and showLogs to pool per 10 seconds
Previously, the showAttrs and showLogs pool the job attributes and logs
per job interval. For example, if the interval is 5 minutes, then the
attributes and/or logs will be refreshed every 5 minutes.

In order to make user can view the latest attributes and/logs
immediately, we changes the interval to 10 seconds.