all: add option to set HTTP method and request type on Job
The HttpMethod field (or http_method in configuration) set the HTTP
method in request.
Its accept only GET, POST, PUT, or DELETE.
This field is optional, default to GET if its empty.

The HttpRequestType field (or http_request_type in configuration) define
the HTTP request type.
Its accept only,

  - query: no header Content-Type to be set, reserved for future use;
  - form: header Content-Type set to "application/x-www-form-urlencoded";
  - json: header Content-Type set to "application/json".

The type "form" and "json" only applicable if the method is POST or PUT.
This field is optional, default to query.