all: send the current epoch on each Job execution
Each Job execution send the parameter named `_karajo_epoch` with value is
current server Unix time.

If the request type is `query` then the parameter is inside the query URL.
If the request type is `form` then the parameter is inside the body.
If the request type is `json` then the parameter is inside the body as
JSON object, for example `{"_karajo_epoch":1656750073}`.