all: implement Hook
Hook is the HTTP endpoint that run a function or list of commands upon
receiving request, a reverse of what a Job.

Each Hook contains Secret for authenticating request, a working directory,
and a callback or list of commands to be executed when the request
received.

The circle is now complete!