all: refactoring Hook logs
The HookLog now store the Hook ID and Status, both in the struct and its
file name using the following format:

  HookID.Counter.Status

By using this pattern, we can know the log number and its status from
the file name.

On the WUI, we display each hook log as number that can be clicked
and open new tab for displaying log only.
This require adding new HTTP API to fetch hook log by its ID and log
number.