all: add option to serve directory to public
In the Environment we add field DirPublic that define a path to serve
to public.

While the WUI is served under "/karajo", a directory dir_public
will be served under "/".
A dir_public can contains sub directory as long as its name is not
"karajo".

In the configuration file, the DirPublic is set under
"karajo::dir_public" option.