all: set the Job and Hook Status before running
The Status is set to "started" so the interface can display different
color.

On Job user interface, if the NextRun is less than now, it will show text
"Running...".

On Hook, set the LastRun to zero time before running, so the WUI can
show status as "Running...".

To test it, we add random sleep on Hooks in testdata.