all: split running the hook into separate goroutine
Previously, hook write the HTTP response after the Call or all of the
Commands are finish.
If the Hook run longer than, say 5 seconds, this may cause the request
that trigger the hook return with timeout.

In this changes, once we receive the request to trigger the Hook and
when the signature is valid, we return with HTTP status 200 immediately
and run the Hook job in the other goroutine.