all: double check for signature header in Hook
In case the Hook use custom header signature, the request that send
by the Job will fail, because no matched header signature found.
This changes make the Hook to double check the signature header with
default HeaderNameXKarajoSign.

While at it, fix Hook status not set to fail if signature is invalid.