all: implement interval based Hook
Previously, Hook can be triggered by sending HTTP POST request to karajo
server.  In most cases we create JobHttp to trigger it, so we need to
define one hook and one JobHttp.

To simplify it, we add an Interval to Hook that works similar to JobHttp
so now we only need to create single Hook.