all: change the API path to execute Job
Previously, the API path to execute Job is "/karajo/job/$job_path".
This may become a conflict in the future (if we want to serve any
information related to job in specific page) and inconsistent API
path.

This changes the API to execute job to "/karajo/api/job/run/$job_path".