all: fix double checking for isPaused
There are two path where Job.execute is called.  One from handleHttp
and one from Start.  The one from handleHttp already check if
job is paused before calling execute.  If we check again inside
execute then that means we doing it twice.

To fix this we move the check to Start method and set the Status as
started before it.