all: simplify logging in JobHttp
Previously to log something to JobHttp we write into Log and mlog.
To simplify it, we register the Log as writer to mlog so we can only
log once.

This changes require un-exporting the Log field and in the UI we fetch
the log only if the HTTP job attributes is expanded.