all: fix possible lock on API environment
Sometimes the request to /karajo/api/environment does not return any
result.  The only explanation is something lock the resource so we cannot
lock it and it will wait forever.

Until we found the cause, remove the lock but expect for data race, but
the data race should be safe since the API only doing read.