all: changes on how the job queued using channel
Previously, a job run using the following flow:

* interval/scheduler timer kicking in
* send the job finish to finished channel

If the job triggered from HTTP request, it will run on its own goroutine.

This changes add third channel, startq, to JobBase that queue the Job.
When the timer kicking in or HTTP request received in it will pushed
to startq.
The startq execute the job and signal the completed job using finishq.