all: change the JobHttp log to use the same as Job
Previously, the JobHttp use single file for log with limited size.
The way it works is quite complex, we need to maintain one file and a
buffer that able to truncate the log if its reached its max size.
We also need to store the job state in separate file.

In this changes, we replace the JobHttp log mechanism to use the same
as Job, where each execution will be log separately.
Not only this give us less complex code, it also remove some duplicate
code.

In the HTTP API we changes the path to get the JobHttp log to match with
Job path.