all: refactoring, rename Job type to JobExec
Using the type named Job in module that contains JobHttp and JobBase
is quite ambigous.
The Job could means the generic one, used by JobBase and JobHttp, while
it the other way around for, the Job type embed the JobBase.

To minimize confusion in the future, we rename the Job type to JobExec
which indicates that the job type is for executing Go code or commands.