all: simplify running and finishing job
In the JobBase we centralize creating new JobLog using newLog, where
job counter incremented and log appending to the job.
If the Job is paused, the jobLog instance will get the same status,
which allow execute method check the status return immediately if its
paused.

The the [JobBase.finish] we centralize pushing the notification
and jobLog to logq.